update abnormality for more accurate duration tracking

+++ TCC.Core/Data/Abnormalities/AbnormalityData.cs
@@ -6,7 +6,7 @@ namespace TCC.Data.Abnormalities
     {
         public uint Id { get; set; }
         public int Stacks { get; set; }
-        public ulong Duration { get; set; }
+        public double Duration { get; set; }
         [JsonIgnore]
         public Abnormality Abnormality => Game.DB.AbnormalityDatabase.Abnormalities[Id];
     }
